--- client/src/App.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-class App extends Component {
-	render() {
-		return (
-			<div className="App">
-				<h1> Hello, World </h1>
-			</div>
-		)
-	}
-}
-
-export default App
