--- client/webpack.config.js
@@ -1,34 +1,0 @@
-const path = require('path')
-const webpack = require('webpack')
-
-module.exports = {
-	entry: './src/index.js',
-	mode: 'development',
-	module: {
-		rules: [
-			{
-				test: /\.(js|jsx)$/,
-				exclude: /(node_modules|bower_components)/,
-				loader: 'babel-loader',
-				options: { presets: ['@babel/env'] },
-			},
-			{
-				test: /\.css$/,
-				use: ['style-loader', 'css-loader'],
-			},
-		],
-	},
-	resolve: { extensions: ['*', '.js', '.jsx'] },
-	output: {
-		path: path.resolve(__dirname, 'dist/'),
-		publicPath: '/dist/',
-		filename: 'bundle.js',
-	},
-	devServer: {
-		proxy: {
-			'/api': 'http://localhost:5501',
-		},
-	},
-
-	plugins: [],
-}
